refactor AIAgentLLMContext
